+++ nf_core/subworkflows/lint/__init__.py
@@ -99,7 +99,7 @@ class SubworkflowLint(ComponentLint):
         """
         # TODO: consider unifying modules and subworkflows lint() function and add it to the ComponentLint class
         # Prompt for subworkflow or all
-        if subworkflow is None and not all_subworkflows:
+        if subworkflow is None and not (local or all_subworkflows):
             questions = [
                 {
                     "type": "list",
@@ -152,7 +152,7 @@ class SubworkflowLint(ComponentLint):
             self.lint_subworkflows(local_subworkflows, registry=registry, local=True)
 
         # Lint nf-core subworkflows
-        if len(remote_subworkflows) > 0:
+        if not local and len(remote_subworkflows) > 0:
             self.lint_subworkflows(remote_subworkflows, registry=registry, local=False)
 
         if print_results:
@@ -208,6 +208,8 @@ class SubworkflowLint(ComponentLint):
         # Only check the main script in case of a local subworkflow
         if local:
             self.main_nf(swf)
+            self.meta_yml(swf, allow_missing=True)
+            self.subworkflow_todos(swf)
             self.passed += [LintResult(swf, *s) for s in swf.passed]
             warned = [LintResult(swf, *m) for m in (swf.warned + swf.failed)]
             if not self.fail_warned: